render sample testbed data to testbeds view
wip

add testbeds menu item to resources dropdown

render testbed names, links, descriptions

rethink/refactor testbeds data
