+++ package.json
@@ -16,7 +16,7 @@
     "gatsby-plugin-sharp": "^2.14.4",
     "gatsby-plugin-styled-components": "^3.10.0",
     "gatsby-plugin-web-font-loader": "^1.0.4",
-    "gatsby-remark-images": "^3.11.1",
+    "gatsby-remark-images": "^5.7.0",
     "gatsby-source-filesystem": "^2.11.1",
     "gatsby-transformer-json": "^2.11.0",
     "gatsby-transformer-remark": "^2.16.1",
