+++ src/data/resources/testbeds.yaml
@@ -1,0 +1,8 @@
+- name: Chameleon Cloud
+  url: https://chameleoncloud.org/
+  description: >
+    lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum 
+- name: CloudLab
+  url: https://cloudlab.com/
+  description: >
+    lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum
