+++ src/hooks/index.js
@@ -9,3 +9,4 @@ export * from './use-solicitations'
 export * from './use-timeline'
 export * from './use-window-width'
 export * from './use-publications'
+export * from './use-testbeds'
